Move the shared sandbox layer to openenv.core.sandbox
The sandbox protocol (SandboxBackend/SandboxHandle/BgJob/ExecResult), the E2B and
Hugging Face backends, the interception proxy, and build_template move from
envs/opencode_env/sandbox/ to src/openenv/core/sandbox/, so both opencode_env and
pi_env import them from core instead of pi_env peer-depending on opencode_env.

- opencode_env and pi_env now import from openenv.core.sandbox; pi_env drops its
  opencode_env peer dependency (the server Dockerfile no longer installs it, and
  `pip install openenv-pi-env` no longer needs opencode_env).
- interception.py is no longer baked into the opencode pre-baked image; the harness
  uploads it at rollout start from core (matching the pi-sandbox image). The opencode
  hf_image dir moves out of sandbox/ to opencode_env/hf_image/ to match pi.
- e2b stays a lazy import so core keeps no hard e2b dependency.
- No backward-compat shim: both opencode_env and the TRL async_grpo path are
  experimental, so external importers (our gists, the TRL opencode example) update to
  openenv.core.sandbox directly.

AI-assisted.
