+++ .github/workflows/docker-build.yml
@@ -110,7 +110,7 @@ jobs:
             dockerfile: envs/opencode_env/server/Dockerfile
             context: envs/opencode_env
           - name: opencode-sandbox
-            dockerfile: envs/opencode_env/sandbox/hf_image/Dockerfile
+            dockerfile: envs/opencode_env/hf_image/Dockerfile
             context: envs/opencode_env
           - name: pi-env
             dockerfile: envs/pi_env/server/Dockerfile
