+++ envs/pi_env/README.md
@@ -76,7 +76,7 @@ turn's `(messages, completion_token_ids, per_token_logps)` to
 
 ### Sandbox backend
 
-`HFSandboxBackend` (from `opencode_env.sandbox`, shared with `opencode_env`)
+`HFSandboxBackend` (from `openenv.core.sandbox`, the shared sandbox layer)
 runs the agent in a Hugging Face sandbox. `image="python:3.12"` cold-installs
 Node 22 (bootstrapped) + the Pi CLI (`npm install -g @mariozechner/pi-coding-agent`)
 + the proxy's Python deps on every rollout. For faster rollouts use the
@@ -88,7 +88,7 @@ sandbox_backend=HFSandboxBackend(image="ghcr.io/huggingface/openenv-pi-sandbox:l
 ```
 
 Any backend satisfying the `SandboxBackend` / `SandboxHandle` / `BgJob`
-protocols in `opencode_env.sandbox.base` can be plugged in the same way.
+protocols in `openenv.core.sandbox.base` can be plugged in the same way.
 
 > The sandbox backend and interception proxy live in `opencode_env` for now;
 > the plan is to consolidate both into `openenv.core` so `pi_env` and
@@ -225,4 +225,4 @@ pi_env/
 ```
 
 The sandbox backend + interception proxy are imported from
-`opencode_env.sandbox`; `pi_env` ships no `sandbox/` of its own.
+`openenv.core.sandbox`; `pi_env` ships no `sandbox/` of its own.
