+++ envs/pi_env/pyproject.toml
@@ -30,10 +30,7 @@ dependencies = [
     "httpx>=0.27.0",
     # >=1.22 ships huggingface_hub.Sandbox (HFSandboxBackend); core only floors 0.20.
     "huggingface_hub>=1.22",
-    # Shared sandbox backend + interception proxy live in opencode_env for now
-    # (plan: consolidate into openenv.core). Kept a peer dep — install from the
-    # same repo (the server Dockerfile installs it explicitly):
-    #   pip install "openenv-opencode-env @ git+https://github.com/huggingface/OpenEnv.git#subdirectory=envs/opencode_env"
+    # Sandbox backend + interception proxy come from openenv.core.sandbox (in `openenv`).
 ]
 
 [project.optional-dependencies]
