+++ src/openenv/core/sandbox/interception.py
