+++ tests/envs/test_opencode_env.py
@@ -305,7 +305,7 @@ def test_start_proxy_keeps_upstream_key_out_of_command() -> None:
 def test_interception_cli_reads_upstream_key_from_env(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    from opencode_env.sandbox import interception
+    from openenv.core.sandbox import interception
 
     captured = {}
